Moved BASE_URL & CALIBER_URL to their own export blocks
