+++ src/environments/environment.ts
@@ -13,11 +13,11 @@ export const environment = {
     "appId": "1:492701958610:web:4a30a1be93803701d3480b",
     "measurementId": "G-DP6XDH9DTW"
   },
-  CALIBER_URL: `https://caliber2-mock.revaturelabs.com/mock`,
-  BASE_URL: `http://ec2-54-237-27-253.compute-1.amazonaws.com:8081/`
+
 };
 
-
+export const CALIBER_URL= `https://caliber2-mock.revaturelabs.com/mock`
+export const BASE_URL= `http://ec2-54-237-27-253.compute-1.amazonaws.com:8081/`
 
 /*
  * For easier debugging in development mode, you can import the following file
